+++ app/src/main/res/values/plurals.xml
@@ -36,4 +36,8 @@
         <item quantity="one">%s patch selected</item>
         <item quantity="other">%s patches selected</item>
     </plurals>
+    <plurals name="home_app_selected">
+        <item quantity="one">%s selected</item>
+        <item quantity="other">%s selected</item>
+    </plurals>
 </resources>
