+++ app/src/main/res/values/strings.xml
@@ -10,8 +10,6 @@
     <string name="home_app_no_hidden">No hidden apps</string>
     <string name="home_app_hidden_apps_title">Hidden apps</string>
     <string name="home_app_hidden_apps_hint">Tap to unhide</string>
-    <string name="home_app_selected_one">One selected</string>
-    <string name="home_app_selected_multiple">Selected: %d</string>
     <string name="android_11_bug_dialog_title">Android 11 bug</string>
     <string name="android_11_bug_dialog_description">The app installation permission must be granted ahead of time to avoid a bug in the Android 11 system that will negatively affect the user experience</string>
